Demonstrate that Times can be implemented w/ Plus
We know this from Haskell, because Data.Semigroup defines `stimes` as
part of the Semigroup typeclass, but doesn't require you to implement
it.

The implementation using Enumerable.Repeat is hardly as efficient as the
individual implementations, but the point here isn't about efficiency,
but about a common, reusable abstraction.

In Haskell, this is addressed by making `stimes` part of the typeclass,
so that implementers can implement a more efficient algorithm than the
default implementation. In C#, the same effect could be achieved by
refactoring `IExpression` to an abstract base class, with `Times` as a
public virtual (overridable) method.